Grip Notice set to False (Potentially remove in the future)
Reference this commit if planning to remove Grip Notice the two script references are listed here. (check that there are no other references). The player.prefab is the other place where the GripNotice exists.

The GripNotice was originally used as a quick way for the player to see the controls and for them to be told that they should teleport to move. We now have a full tutorial to give the same information and also we can show them the menu with the controls if necessary. For those reasons, this object might no longer be necessary and might be removed.