Added Debug Messages for the rest of the Scene
Added debug messages to help design the triggers for the rest of scene 1. Additionally added some wait time in between each of these debug messages (this will be made to match the audio once it is added in). I additionally added new objectives in the objective menu when each objective is called. These names are temporary and will be replaced with the actual objectives as developed.