Added audio/subtitles to Scene 1
Instantiated the Player as well as all of the menus/buttons from the Player prefab, and included a coroutine for waiting for the player's spawn. Using this Player prefab, I was able to add in the audio clips for the entire scene and change the debug commands and timing in order to match the audio. This was successfully tested in the Unity Editor and the audio can be heard with the correct timing.