Fixed Issue Resulting in No Subtitles Playing
Copied in the missing Game Manager game object and the MeshRenderer element of the floor that were missing and causing errors. This resolves the issue with the subtitles not playing, and fixes the menu being open at start, but does not yet fix the issue with bringing up the menu ending the progression of the scene.