Added Final Trigger and Resetting of Endpoints in Objective Two
Added functionality for the EndPointManager to check if the target location (3,1,2) has been reached after every update. When this is reached, the EndPointManager is deactivated, the mass is returned to the origin, and the progression of the scene continues. The player is then prompted to repeat the action and once they do they move on to Objective Three, where the mass is also destroyed.