Implemented Code for Forcing Player to Take Different Endpoint Path
This code forces the player to take a different endpoint path during the endpoint task in Objective 2 that must be completed twice. However, the object still snaps to the endpoint but then immediately snaps back. This interaction will be tested and may be changed, but the base idea will remain. Note: Code will be cleaned up with comments added shortly.