Finished Example Path Construction
Used Endpoint-appearing objects in a pink color to illustrate the possible example path, and created coordinates above the final endpoint that always face the player. I also had to create a material for the Endpoint due to the fact that you cannot change the color of the basic material without creating a new material model.