Updated Coordinate Coloring and Objective Text
Changed Coordinate Coloring above the Desired Location to be in pink parenthesis with the same red, green, and blue coloring as the coordinates above the mass object. Note that this change may be changed later (potentially to white parenthesis). Meanwhile, I also changed the spacing of the Objective Text to leave a gap between each objective.