Fixed Player Move Limit Implementation
Fixed the implementation of the player move limit. Instead of starting over progress is move limit is passed, we now let the player continue to the point, and once they get there, the second go through does not show the path that they first traveled but instead just keeps showing the example path. We also do not impede the player from moving in the same exact path as they did originally, which is something we do if the player does get there within 12 moves.