Fixed Coordinate View and Objective Menu Location
Added background to the coordinates of the mass in order to have them stand out against the axis. Additionally changed them to be near the top of the mass but rotate in a shell-like shape around the mass in order to always point towards the player and always be in front of the mass. Additionally moved objective menu to the other side of the scene in order to get it out of the way