Bugfix: Teleport reticle rotation
Explanation from Discord thread: Basically it assumed that the floor will always be at y-coordinate 0.0f, but with the translation of the floor during the animation, that was no longer true; by changing the 0.0f to the height of the contact point (i.e. the floor's actual y-coordinate), we can ensure that the reticle will always face forward instead of being at an awkward angle!