[Bugfix] Subtitle display on language change
Previously, the background of the subtitle background will show up (but not the text) when changing languages in the menu. Neither should show up.

This is because upon switching languages, the subtitle text needs to update to the new language, which calls a common function that sets the subtitle object (i.e. the background) visibility to true. Note that this common function is also called upon every new line.

I've set it such that it sets it to true only if the game is not paused

However, to account for the case during the tutorial where we do want subtitles to show up while paused, there is the additional check that the subtitleObject's tag is not "Player", which is the case only during that segment of the tutorial.